+++ CHANGELOG.md
@@ -8,6 +8,7 @@
  - add -DBUILD_ZLIB=OFF by default
  - default openCV version build is now 4.5.5
  - OpenCVBuilder.env can be modifyed immediately after his instanciation to enable/disable openCV modules.
+ - add --git-cache or OPENCV_GIT_CACHE env variabe cache git data on local disk
 
 * V 0.4.7
  - add --dry-run parameter to debug openCV build failure.